Linter: Disable govet check-shadowing
check-shadowing is not the default.
We turned it on, but more often than not it's influencing coding style
and practice in a non-idiomatic way, e.g. when using a local err or ctx

It's also painful when you want to deliberately shadow.
We could use no-lint, but it becomes overbearing.